* Ihotelreport.userprefs: Completed Version
* Help.xaml:
* tabbar.cs:
* TabAgent.cs:
* tabagency.cs:
* Aboutpage.xaml:
* Info.plist:
* AgencyRoom.xaml:
* Magency.xaml.cs:
* tabpieagency.cs:
* Yagency.xaml.cs:
* MTD_Revenue.xaml:
* YTD_Revenue.xaml:
* iTunesArtwork:
* Aboutpage.xaml.cs:
* RootPage.cs:
* Staticreport.xaml:
* Y_Compare.xaml.cs:
* Revenueagency.xaml:
* Revenuereport.xaml:
* MTD_Revenue.xaml.cs:
* Ihotelreport.csproj:
* AreaHeaderView.xaml:
* YTD_Revenue.xaml.cs:
* iTunesArtwork@2x:
* MenuListData.cs:
* Revenuereport.xaml.cs:
* bed.png:
* AreaPageyearnight.xaml:
* AreaPageyearevenue.xaml:
* LoginPage.xaml.cs:
* night.png:
* AreaPagetodaynight.xaml:
* AreaPagemonthnight.xaml:
* AreaPageyearnight.xaml.cs:
* AreaPagemonthrevenue.xaml:
* revenue.png:
* AreaPagetodayrevenue.xaml:
* Ihotelreport.dll:
* Ihotelreport.pdb:
* Model0.cs:
* AreaPagetodaynight.xaml.cs:
* AreaPageyearevenue.xaml.cs:
* Ihotelreport.dll:
* AreaPagemonthnight.xaml.cs:
* Ihotelreport.pdb:
* Ihotelreport.iOS.csproj:
* AreaPagemonthrevenue.xaml.cs:
* AreaPagetodayrevenue.xaml.cs:
* Model.cs:
* HelveticaNeue Thin.ttf:
* ColumnChart0.xaml:
* ColumnChart1.xaml:
* ColumnChart2.xaml:
* ColumnChart0.xaml.cs:
* ColumnChart1.xaml.cs:
* ColumnChart2.xaml.cs:
* ColumnChart.xaml:
* Resource.Designer.cs:
* tabbar.xml:
* toolbar.xml:
* ColumnSeriesViewModel2.cs:
* ColumnSeriesViewModel1.cs:
* ColumnSeriesViewModel0.cs:
* 58.png:
* 76.png:
* 80.png:
* 29.png:
* 40.png:
* 120.png:
* 152.png:
* Ihotelreport.AreaPageyearnight.xaml.g.cs:
* Ihotelreport.csproj.FileListAbsolute.txt:
* Ihotelreport.AreaPagemonthnight.xaml.g.cs:
* Ihotelreport.AreaPagetodaynight.xaml.g.cs:
* Ihotelreport.AreaPageyearevenue.xaml.g.cs:
* Ihotelreport.AreaPagetodayrevenue.xaml.g.cs:
* Ihotelreport.AreaPagemonthrevenue.xaml.g.cs:
* Ihotelreport.csprojResolveAssemblyReference.cache:
* schemas.android.com.apk.res.android.xsd:
* treemap_layout.xml:
* sfautocomplete_layout.xml:

+++ Ihotelreport/Ihotelreport/Ihotelreport/Y_Compare.xaml.cs
@@ -18,8 +18,8 @@ namespace Ihotelreport
 		CultureInfo UsaCulture = new CultureInfo("en-US");
 		string[] items;
 		string[] items2;
-		float[] totalold;
-		float[] totalnew;
+		double[] totalold;
+        double[] totalnew;
 		float[] totalSum;
 		string[] agencyName;
 		Double[] agencyTotal;
@@ -128,8 +128,8 @@ namespace Ihotelreport
 				int i = 0;
 				items = new string[count];
 				id = new string[count];
-				totalnew = new float[count];
-				totalold = new float[count];
+				totalnew = new double[count];
+				totalold = new double[count];
 				foreach (var aaa in Items.dataResult)
 				{
 					Debug.WriteLine("Name: " + aaa.name + " ID = " + aaa.id + " PresentSum = " + aaa.presentSum + " PreviousSum = " + aaa.presentSum);
@@ -142,8 +142,8 @@ namespace Ihotelreport
                     {
                         items[i] = aaa.name;
                         id[i] = aaa.id;
-                        totalnew[i] = float.Parse(aaa.presentSum);
-                        totalold[i] = float.Parse(aaa.previousSum);
+                        totalnew[i] = double.Parse(aaa.presentSum);
+                        totalold[i] = double.Parse(aaa.previousSum);
                         i++;
                     }
 				}
